--- AVOS/AVOSCloudTests/Swift/AVInstallation_TestCase.swift
@@ -1,13 +1,0 @@
-//
-//  AVInstallation_TestCase.swift
-//  AVOS
-//
-//  Created by zapcannon87 on 05/01/2018.
-//  Copyright © 2018 LeanCloud Inc. All rights reserved.
-//
-
-import XCTest
-
-class AVInstallation_TestCase: LCTestBase {
-    
-}
