+++ AVOS/TestsBase/LCTestBase.swift
@@ -10,19 +10,6 @@ import XCTest
 
 class LCTestBase: XCTestCase {
     
-    enum TestApp {
-        case CN_North
-        case CN_East
-        case US
-        var appInfo: (id: String, key: String) {
-            switch self {
-            case .CN_North: return (id: "S5vDI3IeCk1NLLiM1aFg3262-gzGzoHsz", key: "7g5pPsI55piz2PRLPWK5MPz0")
-            case .CN_East: return (id: "uwWkfssEBRtrxVpQWEnFtqfr-9Nh9j0Va", key: "9OaLpoW21lIQtRYzJya4WHUR")
-            case .US: return (id: "eX7urCufwLd6X5mHxt7V12nL-MdYXbMMI", key: "PrmzHPnRXjXezS54KryuHMG6")
-            }
-        }
-    }
-    
     override class func setUp() {
         super.setUp()
         
@@ -33,9 +20,9 @@ class LCTestBase: XCTestCase {
         
         /// set region
         if let region: String = env.app_REGION {
-            switch region {
-            case "us": AVOSCloud.setServiceRegion(.US)
-            case "cn": AVOSCloud.setServiceRegion(.CN)
+            switch region.uppercased() {
+            case "US": AVOSCloud.setServiceRegion(.US)
+            case "CN": AVOSCloud.setServiceRegion(.CN)
             default: AVOSCloud.setServiceRegion(.CN)
             }
         } else {
@@ -54,8 +41,8 @@ class LCTestBase: XCTestCase {
         if let appId: String = env.app_ID, let appKey: String = env.app_KEY {
             AVOSCloud.setApplicationId(appId, clientKey: appKey)
         } else {
-            let testRegion: TestApp = .CN_North
-            AVOSCloud.setApplicationId(testRegion.appInfo.id, clientKey: testRegion.appInfo.key)
+            let testApp: TestApp = .ChinaNorth
+            AVOSCloud.setApplicationId(testApp.appInfo.id, clientKey: testApp.appInfo.key)
         }
         
         AVOSCloud.setAllLogsEnabled(true)
@@ -71,6 +58,27 @@ class LCTestBase: XCTestCase {
     
     override func tearDown() {
         super.tearDown()
+    }
+    
+}
+
+extension LCTestBase {
+    
+    var isServerTesting: Bool {
+        return LCTestEnvironment.sharedInstance().isServerTesting
+    }
+    
+    enum TestApp {
+        case ChinaNorth
+        case ChinaEast
+        case US
+        var appInfo: (id: String, key: String) {
+            switch self {
+            case .ChinaNorth: return (id: "S5vDI3IeCk1NLLiM1aFg3262-gzGzoHsz", key: "7g5pPsI55piz2PRLPWK5MPz0")
+            case .ChinaEast: return (id: "uwWkfssEBRtrxVpQWEnFtqfr-9Nh9j0Va", key: "9OaLpoW21lIQtRYzJya4WHUR")
+            case .US: return (id: "eX7urCufwLd6X5mHxt7V12nL-MdYXbMMI", key: "PrmzHPnRXjXezS54KryuHMG6")
+            }
+        }
     }
     
 }
